Updated the expected results for the "bacteria_run_linux" shinytests.
NOTE 1: The last commit message (commit
88e5c3073ae221ebf1ca082d9d915648b0e170df) should read 'Updated the
expected results for the "bacteria_run_windows" shinytests.' instead of
'Updated the expected results for the "bacteria_run_[...]" shinytests.'
NOTE 2: The reason for removing the internal links to the tab "MCMC
diagnostics" from the notifications thrown after the Stan run was to
avoid that "posterior_navlist_ID" is automatically set to "MCMC
diagnostics" (since the creation of the notification creates the
internal link (e.g. "mcmc_link2") and thereby activates it for the first
time). Up to now, this automatic activation of the internal link did not
cause any problems, but it might cause undesired behavior in the future.