Revert "Catch warnings and errors from the `brms::conditional_effects()` call."
This reverts commit 7338bf63faae45a01fdb4ad97c81a85e0381516d. Reason: This might
face the user with error messages
"All specified effects are invalid for this model. [...]"
which should usually only be shown temporarily, though. And if such error messages
are shown permanently (not only temporarily), then they really indicate a problem, so
they should be shown in that case. Thus, simply suppressing the errors is not an option.
The most desirable option would be to freeze the `input$term_sel` object before updating
it, but this is an issue which is not completely resolved yet in **shiny** (see the same
problem with `C_prior_default`).