Re-introduce (and modify) some `brms::get_prior()` code removed by
commit 055d0b2ca2ae298d1a4b449dd3d5152dd9b24494 ("
When retrieving the default prior: Return empty prior when brms::get_prior() leads to a "try-error". This replaces the old checks performed at the beginning of the code for "C_prior_default" and should be less error-prone.
") which is necessary in conjunction with
commit 29f86011e9e1c5c5e3a8d9bfc220bb6d66c1983c ("
Always show the errors from `brms::get_prior()` to the user (to avoid having to

assert that offset variables are numeric (and show other possibly important
errors, too)).
") to avoid flooding the user with harmless error messages when switching (or
clearing) the dataset.